requirements.txt
pandas>=2.0
numpy>=1.25
scikit-learn>=1.3
shap>=0.44
rdkit-pypi>=2022.9.5
matplotlib>=3.7
seaborn>=0.12
streamlit>=1.36
pytest>=8.0
